+++ pages/cifras_proyectos_periodo.py
@@ -63,7 +63,7 @@ def total_function(facultad, anio, dataframe):
 
 data_2["Año"] = data_2["Año"].astype('str')
 data_2.fillna(0, inplace=True)
-data_2['cifra'] = data_2['cifra'].astype('int')
+data_2['cifra'] = data_2['cifra'].astype('float')
 
 data_2.apply(lambda x: total_function(x['Facultad'], x['Año'], data_2), axis=1)
 total_data_2 = data_2['cifra'].sum()
@@ -72,7 +72,7 @@ total_data_2 = data_2['cifra'].sum()
 
 data_3["Año"] = data_3["Año"].astype('str')
 data_3.fillna(0, inplace=True)
-data_3['cifra'] = data_3['cifra'].astype('int')
+data_3['cifra'] = data_3['cifra'].astype('float')
 
 data_3.apply(lambda x: total_function(x['Facultad'], x['Año'], data_3), axis=1)
 total_data_3 = data_3['cifra'].sum()
@@ -81,7 +81,7 @@ total_data_3 = data_3['cifra'].sum()
 
 data_4["Año"] = data_4["Año"].astype('str')
 data_4.fillna(0, inplace=True)
-data_4['cifra'] = data_4['cifra'].astype('int')
+data_4['cifra'] = data_4['cifra'].astype('float')
 
 data_4.apply(lambda x: total_function(x['Facultad'], x['Año'], data_4), axis=1)
 total_data_4 = data_4['cifra'].sum()
